server: GET /users, /users/<id>
Project-structure
-----------------

These are the first lines of go and we are quite deep in the directory
structure. We may find some Balrog here.

Web
---

At first attempt I try to build the server from the standard library.
If  fail mux router could be tried, based on the advice of:
 [blog]: https://thenewstack.io/make-a-restful-json-api-go/
It would be nice if htt.handleFunc could match method, too.

However the ResponseWriter abstraction gives a flexible connection to
any manipulation (like json parsing)

Log
---

Log level would be nice from a mature built-in library, that can be a
deal breaker in my attempt to use only built-in libraries.